+++ OneBlink.SDK/models/FormSubmissionEventConfiguration.cs
@@ -2,11 +2,6 @@ using System;
 using System.Collections.Generic;
 namespace OneBlink.SDK.Model
 {
-
-    //Being kept right now due to backwards compatability, will be deleted upon next breaking change.
-    public class FormSubmissionEventConfigration : FormSubmissionEventConfiguration
-    {
-    }
     public class FormSubmissionEventConfiguration : PDFConfiguration
     {
         public string url
@@ -45,6 +40,7 @@ namespace OneBlink.SDK.Model
         {
             get; set;
         }
+        [Obsolete("use toEmail instead")]
         public string email
         {
             get; set;
@@ -142,6 +138,18 @@ namespace OneBlink.SDK.Model
             get; set;
         }
         public FormSubmissionEventEmailTemplate emailTemplate
+        {
+            get; set;
+        }
+        public List<string> toEmail
+        {
+            get; set;
+        }
+        public List<string> ccEmail
+        {
+            get; set;
+        }
+        public List<string> bccEmail
         {
             get; set;
         }