+++ CHANGELOG.md
@@ -6,8 +6,18 @@
 
 -   `hintPosition` to `FormElement`
 -   `postSubmissionReceipt` to `Form`
--   FormExternalIdGeneration model
--   ReceiptConfiguration model
+-   `FormExternalIdGeneration` model
+-   `ReceiptConfiguration` model
+-   `toEmail`, `ccEmail` and `bccEmail` to `FormSubmissionEventConfiguration`
+
+### Changed
+
+-   Deprecated `email` from `FormSubmissionEventConfiguration`
+
+### Removed
+
+-   **[BREAKING]** Removed `FormSubmissionEventConfigration`
+-   **[BREAKING]** Removed `email` from `FormSubmissionEvent.CreatePDFSubmissionEvent()` and `FormSubmissionEvent.CreateEmailSubmissionEvent()`
 
 ## 4.1.1 (2023-03-13)